+++ CHANGELOG.md
@@ -13,6 +13,7 @@ facing release notes.
 ## v2.19.0
 
   *  [Misc] RBA-135: Changed 'fact synthesised' wording to to read 'Condition not met'.
+  *  [Fix]  RB-1460: Evidence Tree is not rendering text correctly.
 
 ## v2.18.0